microbit-quiz: Created yotta module around code
