+++ .gitignore
@@ -1,0 +1,3 @@
+build
+yotta_targets
+yotta_modules
