+++ source/MicroBitQuiz.h
@@ -1,0 +1,7 @@
+#ifndef MICROBIT_QUIZ_H
+#define MICORBIT_QUIZ_H
+
+//#define MICROBIT_SAMPLE_VOTER_STUDENT
+#define MICROBIT_SAMPLE_VOTER_TEACHER
+
+#endif
